backend/requirements.txt
absl-py==0.11.0
aiofiles==0.6.0
aniso8601==8.1.0
appdirs==1.4.4
astor==0.8.1
attrs==20.3.0
audioread==2.1.9
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dnspython==2.1.0
fastapi==0.63.0
fastapi-contrib==0.2.9
Flask==1.1.2
flask-restplus==0.13.0
gast==0.4.0
gin-config==0.4.0
grpcio==1.34.0
gunicorn==20.0.4
h11==0.12.0
h5py==3.1.0
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.13.2
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.7.1
llvmlite==0.31.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==4.0.3
motor==2.3.0
numba==0.48.0
numpy==1.16.4
packaging==20.8
pooch==1.3.0
protobuf==3.14.0
pysndfx==0.3.6
pycparser==2.20
pydantic==1.7.3
pydub==0.24.1
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
resampy==0.2.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
sox==1.4.1
starlette==0.13.6
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.31.1
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
uvicorn==0.13.3
Werkzeug==0.16.1
zipp==3.4.0
python-multipart==0.0.5
uvloop==0.15.2
httptools==0.2.0